requirements/core.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
pydantic
pydantic-settings
python-dotenv
aiosqlite
python-jose
